requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
dbt-postgres
astronomer-cosmos
pendulum
python-dotenv
apache-airflow-providers-airbyte